% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_conn.R
\name{arrow_con}
\alias{arrow_con}
\title{Connection to Arrow Disk Dataset}
\usage{
arrow_con(parquet_path)
}
\arguments{
\item{parquet_path}{}
}
\value{
arrow connection
}
\description{
Connection to Arrow Disk Dataset
}
