% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_address}
\alias{validate_address}
\title{Validate address with USPS}
\usage{
validate_address(address)
}
\arguments{
\item{address}{a standard US address}
}
\value{
tibble with validation
}
\description{
Validate address with USPS
}
\examples{
address <- "1600 Pennsylvania Avenue NW, Washington, DC 20500"

#' validate_format(tf)

}
