% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_format}
\alias{validate_format}
\title{Validate file format}
\usage{
validate_format(df, req_cols = c("address", "id"), opt_cols = c("lat", "lng"))
}
\arguments{
\item{df}{}

\item{req_cols}{}

\item{opt_cols}{}
}
\value{
df
}
\description{
Validate file format
}
\examples{
tdf <- data.frame(
  id = "123",
  address = "1600 Pennsylvania Avenue NW, Washington, DC 20500",
  something = "something"
)

validate_format(tf)
}
